combine common code
