+++ algorithms/sorts/Makefile
@@ -3,12 +3,16 @@ CFLAGS	= -Ofast -Wall -std=c99
 
 SRCS	:= $(wildcard *.c)
 HDRS	:= $(wildcard *.h)
+OBJS	:= $(wildcard *.o)
 TGTS	:= $(patsubst %.c,%,$(SRCS))
 
-all: $(TGTS)
+all: sortutils $(TGTS)
 
 $(TGTS): $(SRCS)
-	$(CC) $(CFLAGS) $< -o $@
+	$(CC) $(CFLAGS) $< sortutils.o -o $@
+
+sortutils:
+	$(CC) $(CFLAGS) -c sortutils.c -o $@.o
 
 clean:
-	rm -rf $(TGTS) a.out
+	rm -rf $(TGTS) $(OBJS) a.out *.txt
