+++ algorithms/sorts/sortutils.h
@@ -1,0 +1,8 @@
+#define MAXLINE 1024
+#define MAXSIZE 65536
+
+typedef char *str;
+
+int read(str strings[]); // return the number of lines
+
+void write(str strings[]);
